@@ -33,7 +33,9 @@
     "rehype-katex": "^7.0.1",
     "rehype-prism-plus": "^2.0.0",
     "rehype-slug": "^6.0.0",
+    "rehype-stringify": "^10.0.0",
     "remark-gfm": "^4.0.0",
+    "remark-rehype": "^11.1.0",
     "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7",
     "unist-util-visit": "^5.0.0"
@@ -50,6 +52,6 @@
     "postcss": "^8",
     "tailwindcss": "^3.4.10",
     "ts-node": "^10.9.2",
-    "typescript": "^5"
+    "typescript": "^5.6.2"
   }
 }
