@@ -43,22 +43,12 @@ type BaseMdxFrontmatter = {
 export async function getDocument(slug: string) {
   try {
     const contentPath = getDocumentPath(slug);
-    const rawMdx = await fs.readFile(contentPath, 'utf-8');
-    const parsedMdx = await parseMdx<BaseMdxFrontmatter>(rawMdx);
-
-    return {
-      frontmatter: parsedMdx.frontmatter,
-      content: rawMdx,
-    };
+    const rawMdx = await fs.readFile(contentPath, "utf-8");
+    return await parseMdx<BaseMdxFrontmatter>(rawMdx);
   } catch (err) {
     console.log(err);
     return null;
   }
-}
-
-
-function getDocumentPath(slug: string) {
-  return path.join(process.cwd(), "/contents/docs/", `${slug}/index.mdx`);
 }
 
 export function getPreviousNext(path: string) {
@@ -67,6 +57,10 @@ export function getPreviousNext(path: string) {
     prev: PageRoutes[index - 1],
     next: PageRoutes[index + 1],
   };
+}
+
+function getDocumentPath(slug: string) {
+  return path.join(process.cwd(), "/contents/docs/", `${slug}/index.mdx`);
 }
 
 const preCopy = () => (tree: any) => {
